requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
google-auth>=2.20.0
google-api-python-client>=2.90.0
python-dotenv>=1.0.0
